backend/requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
altgraph
annotated-types
anyio
asgiref
attrs
backoff
bcrypt
beautifulsoup4
black
blinker
blis
build
cachetools
catalogue
certifi
cffi
cfgv
chardet
charset-normalizer
chroma-hnswlib
chromadb
click
cloudpathlib
colorama
coloredlogs
confection
contourpy
coverage
cryptography
cssselect
cycler
cymem
dataclasses-json
deprecation
distlib
distro
dnspython
durationpy
ecdsa
email_validator
fastapi
filelock
flake8
Flask
flatbuffers
fonttools
frozenlist
fsspec
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
groq
gotrue
greenlet
grpcio
h11
h2
hpack
httpcore
httplib2
httptools
httpx
huggingface-hub
humanfriendly
hvac
hyperframe
identify
idna
importlib_metadata
importlib_resources
iniconfig
itsdangerous
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langcodes
language_data
ldap3
loguru
lxml
marisa-trie
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mccabe
mdurl
ml_dtypes
mmh3
mpmath
multidict
murmurhash
mypy
mypy_extensions
netifaces
networkx
nltk
nodeenv
numpy
oauthlib
onnxruntime
openai
opencv-contrib-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
opt_einsum
orjson
overrides
packaging
passlib
pathlib2
pathspec
pdfminer.six
pdfplumber
pefile
pillow
platformdirs
pluggy
postgrest
posthog
pre-commit
preshed
propcache
proto-plus
protobuf
psutil
psycopg2-binary
pulsar-client
pyasn1
pyasn1_modules
pybase64
pycodestyle
pycparser
pydantic
pydantic-settings
pydantic_core
pyflakes
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyJWT
pyotp
pyparsing
PyPDF2
pypdfium2
pyphen
PyPika
pyproject_hooks
pyreadline3
pyspnego
pytest
pytest-asyncio
pytest-cov
python-dateutil
python-docx
python-dotenv
python-jose
python-json-logger
python-multipart
python-pptx
pywin32
pywin32-ctypes
PyYAML
qrcode
readability-lxml
realtime
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
schedule
scikit-learn
scipy
sentence-transformers
sentencepiece
sentry-sdk
shellingham
six
smart-open
sniffio
sounddevice
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
sspilib
starlette
storage3
StrEnum
structlog
supabase
supabase-auth
supabase-functions
supafunc
sympy
tenacity
textstat
thinc
threadpoolctl
tiktoken
tokenizers
torch
torchvision
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
vaderSentiment
virtualenv
wasabi
watchfiles
weasel
websocket-client
websockets
Werkzeug
win32_setctime
wrapt
xlsxwriter
yarl
zipp
